// src/app/models/marketing-management/notes/offer-price-model.ts
export class OfferPriceModel {
   _id : string;
   fullName : string;
   email : string;
   phone : string;
   address : string;
   service: string;
   unit: string;
   packageMaterial: string;
   weight : number;
   length: number;
   width: number;
   height: number;
   departurePlace: string;
   destinationPlace: string;
   timeTarget: number;
   temperatureRequired: string;
   priceRequired: number;
   wareContent: string;
   note: string;
   status: number;
}